chore: move shared components into separate folder
