+++ src/components/Homepage/Homepage.tsx
@@ -4,7 +4,7 @@ import bgBannerMobile from '@/assets/image-header.jpg'
 import bgFeaturedTwoDesktop from '@/assets/zx7-desktop.jpg'
 import bgFeaturedTwoMobile from '@/assets/zx7-mobile.jpg'
 import BannerSection from './subcomponents/BannerSection'
-import CategoriesSection from './subcomponents/CategoriesSection'
+import CategoriesSection from '../shared/CategoriesSection'
 import FeaturedSection from './subcomponents/FeaturedSection'
 import styles from './Homepage.module.sass'
