+++ src/components/shared/CategoriesSection.tsx
